Factored preview-to-code mapping into a separate module.
The LivePreview class is now its own event emitter, which emits a
refresh event every time the preview content changes. This can be used by
other classes like the preview-to-code mapper to provide their functionality.
