+++ js/main.js
@@ -49,6 +49,7 @@ define("main", function(require) {
       ContextSensitiveHelp = require("fc/ui/context-sensitive-help"),
       ErrorHelp = require("fc/ui/error-help"),
       LivePreview = require("fc/ui/live-preview"),
+      PreviewToEditorMapping = require("fc/ui/preview-to-editor-mapping"),
       HistoryUI = require("fc/ui/history"),
       PublishUI = require("fc/ui/publish"),
       ShareUI = require("fc/ui/share"),
@@ -84,6 +85,7 @@ define("main", function(require) {
     ignoreErrors: true,
     previewArea: AppReady.previewArea
   });
+  var previewToEditorMapping = PreviewToEditorMapping(preview);
   var publisher = Publisher(publishURL);
   var publishUI = PublishUI({
     codeMirror: codeMirror,
