+++ test/index.html
@@ -35,8 +35,9 @@
       "test/test-indexable-codemirror",
       "test/test-parsing-codemirror",
       "test/test-mark-tracker",
-      "test/live-preview/test-live-preview",
-      "test/publisher/test-publisher"
+      "test/test-live-preview",
+      "test/publisher/test-publisher",
+      "test/preview-to-editor-mapping/test-preview-to-editor-mapping"
     ], function() {
       if (QUnit.config.blocking)
         QUnit.config.autostart = true;
